% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Function retrying to parse data from Binance API}
\usage{
get_bitcoin_price(retried = 0)
}
\description{
Function retrying to parse data from Binance API
}
