% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hist_FX}
\alias{hist_FX}
\title{Querying historical exchange rates}
\usage{
hist_FX(from, to, num_days)
}
\arguments{
\item{from}{specify ccy code for fx to convert from}

\item{to}{specify ccy code for fx to convert to}

\item{num_days}{specify number of days to go back in time - returned days may be less due to non-business dates}
}
\description{
Specify the FXs and the days to go back in time
}
