% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_bitcoin_price}
\alias{get_bitcoin_price}
\title{Hitting the Binance API to get the most recent price of Bitcoin in USD}
\usage{
get_bitcoin_price(retried = 0)
}
\arguments{
\item{retried}{Number of retries in the previous step}
}
\description{
This is a wrapper around \code{binancer} package...
}
